Move ImprovedMySiteFragment.kt to mysite package
